feat: 🎸 refactor file structure google analytics
