+++ app/features/google-analytics/constants/tracking-id.ts
@@ -1,1 +1,2 @@
-export const trackingId = import.meta.env.VITE_GOOGLE_ANALYTIC_TRACKING_ID;
+export const trackingId: string = import.meta.env
+  .VITE_GOOGLE_ANALYTIC_TRACKING_ID;
