+++ app/features/google-analytics/hooks/use-page-view.ts
@@ -3,12 +3,6 @@ import { useLocation } from "react-router";
 
 import { trackingId } from "../constants/tracking-id";
 
-declare global {
-  interface Window {
-    gtag?: (...args: any[]) => void;
-  }
-}
-
 /**
  * Custom React hook that sends a Google Analytics page view event
  * whenever the current location changes.
@@ -21,7 +15,13 @@ export function usePageView() {
   const location = useLocation();
 
   useEffect(() => {
-    if (!import.meta.env.PROD || !window.gtag || !trackingId) return;
+    if (
+      !import.meta.env.PROD ||
+      typeof window.gtag !== "function" ||
+      !trackingId
+    ) {
+      return;
+    }
 
     window.gtag("event", "page_view", {
       page_path: location.pathname + location.search,
