+++ app/features/google-analytics/index.ts
@@ -1,0 +1,5 @@
+import { HeadScriptTrackingTag } from "./components/head-script-tracking-tag";
+import { usePageView } from "./hooks/use-page-view";
+import { trackCustomEvents } from "./utils/track-custom-events";
+
+export { HeadScriptTrackingTag, usePageView, trackCustomEvents };
