+++ app/features/google-analytics/types/custom-global-window.d.ts
@@ -1,0 +1,7 @@
+import type { GTagFunction } from "./gtag-function";
+
+export declare global {
+  interface Window {
+    gtag: GTagFunction;
+  }
+}
