+++ app/features/google-analytics/types/gtag-function.ts
@@ -1,0 +1,1 @@
+export type GTagFunction = (...args: any[]) => void;
