+++ app/features/google-analytics/types/tracking-event.ts
@@ -1,0 +1,7 @@
+export type TrackingEvent = {
+  event_category: string;
+  event_label: string;
+  event_action?: string;
+  event_value?: string;
+  [key: string]: string | number | undefined;
+};
