+++ app/features/landing-page/components/donate-section.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from "@kotonosora/i18n-react";
 import { Button } from "@kotonosora/ui/components/ui/button";
 import { Link } from "react-router";
 
-import { trackCustomEvents } from "~/features/google-analytics/utils/track-custom-events";
+import { trackCustomEvents } from "~/features/google-analytics";
 
 export function DonateSection() {
   const t = useTranslation();
