+++ app/features/shared/components/github-button.tsx
@@ -4,7 +4,7 @@ import { Link, useLoaderData } from "react-router";
 
 import type { GeneralInformationType } from "../types/type";
 
-import { trackCustomEvents } from "~/features/google-analytics/utils/track-custom-events";
+import { trackCustomEvents } from "~/features/google-analytics";
 
 import GitHubLogoDark from "../assets/github-invertocat-dark.svg?no-inline";
 import GitHubLogoLight from "../assets/github-invertocat-light.svg?no-inline";
