+++ app/features/shared/header-navigation/components/mode-switcher.tsx
@@ -2,7 +2,7 @@ import { Button } from "@kotonosora/ui/components/ui/button";
 import { MoonIcon, SunIcon } from "lucide-react";
 import { Theme, useTheme } from "remix-themes";
 
-import { trackCustomEvents } from "~/features/google-analytics/utils/track-custom-events";
+import { trackCustomEvents } from "~/features/google-analytics";
 
 export function getNextTheme(theme: Theme | null) {
   return theme === Theme.DARK ? Theme.LIGHT : Theme.DARK;
